# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hack MAX API",
    "description": "API бота «Умный дом» для MAX: обращения жителей по проблемам дома, их классификация и переписка с управляющей компанией.",
    "version": "0.1.0"
  },
  "paths": {
    "/users/demo": {
      "post": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Создать демо-пользователя",
        "description": "Создаёт демо-пользователя со случайными домами, если он ещё не существует. Возвращает данные пользователя.",
        "operationId": "demo_create_user_users_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoUserIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appeals/create": {
      "post": {
        "tags": [
          "Обращения"
        ],
        "summary": "Создать обращение",
        "description": "Создаёт новое обращение жителя. Классифицирует текст через LLM, отправляет уведомление на email. Возвращает детальную информацию об обращении.",
        "operationId": "create_appeal_appeals_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealDetailedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appeals/{appeal_id}/update": {
      "patch": {
        "tags": [
          "Обращения"
        ],
        "summary": "Обновить статус обращения",
        "description": "Изменяет статус обращения. При изменении статуса отправляет уведомление в чат MAX и на email.",
        "operationId": "update_appeal_status_appeals__appeal_id__update_patch",
        "parameters": [
          {
            "name": "appeal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appeal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appeals/{appeal_id}/messages": {
      "get": {
        "tags": [
          "Обращения"
        ],
        "summary": "Получить обращение по ID",
        "description": "Возвращает детальную информацию об обращении по его ID, включая сообщения.",
        "operationId": "get_appeal_appeals__appeal_id__messages_get",
        "parameters": [
          {
            "name": "appeal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appeal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealDetailedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/houses/{house_id}/appeals": {
      "get": {
        "tags": [
          "Обращения"
        ],
        "summary": "Список обращений по дому",
        "description": "Возвращает список всех обращений, связанных с указанным домом.",
        "operationId": "list_appeals_by_house_id_houses__house_id__appeals_get",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppealOut"
                  },
                  "title": "Response List Appeals By House Id Houses  House Id  Appeals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appeals/{appeal_id}/message": {
      "post": {
        "tags": [
          "Обращения"
        ],
        "summary": "Отправить сообщение в обращение",
        "description": "Отправляет дополнительное сообщение в существующее обращение.Текст проходит классификацию через LLM, результат отправляется на email.",
        "operationId": "send_message_appeal_appeals__appeal_id__message_post",
        "parameters": [
          {
            "name": "appeal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appeal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/houses": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Список домов пользователя",
        "description": "Возвращает список домов, привязанных к указанному пользователю.",
        "operationId": "list_user_houses_users__user_id__houses_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseOut"
                  },
                  "title": "Response List User Houses Users  User Id  Houses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Информация о пользователе",
        "description": "Возвращает полную информацию о пользователе, включая привязанные дома и историю обращений.",
        "operationId": "get_user_info_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/appeals": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Все обращения пользователя (детально)",
        "description": "Возвращает список всех обращений пользователя с детальной информацией, включая сообщения.",
        "operationId": "list_user_appeals_users__user_id__appeals_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppealDetailedOut"
                  },
                  "title": "Response List User Appeals Users  User Id  Appeals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/appeals/short": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Все обращения пользователя (кратко)",
        "description": "Возвращает краткий список обращений пользователя без детальной информации о сообщениях.",
        "operationId": "list_user_appeals_short_users__user_id__appeals_short_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppealOut"
                  },
                  "title": "Response List User Appeals Short Users  User Id  Appeals Short Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/houses": {
      "get": {
        "tags": [
          "Дома"
        ],
        "summary": "Список всех домов",
        "description": "Возвращает список всех домов в базе данных.",
        "operationId": "list_houses_houses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HouseOut"
                  },
                  "type": "array",
                  "title": "Response List Houses Houses Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Дома"
        ],
        "summary": "Добавить дом",
        "description": "Добавляет новый дом в базу данных. Возвращает 400, если дом с таким адресом уже существует.",
        "operationId": "add_house_houses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/houses/{house_id}/messages": {
      "get": {
        "tags": [
          "Дома"
        ],
        "summary": "Информация о доме с сообщениями",
        "description": "Возвращает детальную информацию о доме, включая историю сообщений общего чата.",
        "operationId": "get_house_info_houses__house_id__messages_get",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/houses/{house_id}/message": {
      "post": {
        "tags": [
          "Дома"
        ],
        "summary": "Отправить сообщение в общий чат дома",
        "description": "Отправляет сообщение пользователя в общий чат дома (не связано с конкретным обращением).",
        "operationId": "send_message_houses__house_id__message_post",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddressIn": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "AddressIn"
      },
...
      "DemoUserIn": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "chat_id": {
            "type": "integer",
            "title": "Chat Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "chat_id",
          "name"
        ],
        "title": "DemoUserIn"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "max_chat_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Chat Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "houses": {
            "items": {
              "$ref": "#/components/schemas/HouseOut"
            },
            "type": "array",
            "title": "Houses",
            "default": []
          },
          "appeals": {
            "items": {
              "$ref": "#/components/schemas/AppealOut"
            },
            "type": "array",
            "title": "Appeals",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "UserOut"
      },
...
    }
  },
...
  "servers": [
    {
      "url": "https://domovoy.stirkk.ru/api",
      "description": "Прод"
    },
    {
      "url": "http://localhost:8000",
      "description": "Локальный Docker"
    }
  ]
}